requirements.txt
fastapi[all]
uvicorn
python-decouple
pymongo
selenium
webdriver-manager
pytesseract